backend/requirements.txt
Django==3.2.13
djangorestframework==3.14.0
django-filter==21.1
mysqlclient==2.0.3
mysql-connector-python==8.0.33
pyyaml==6.0
django-cors-headers==3.10.0
djangorestframework-simplejwt==5.3.0
stripe==5.5.0
